requirements.txt
streamlit
st-supabase-connection
gotrue
google-generativeai
